requirements.txt
Flask==2.3.2
flask-cors==3.0.10
werkzeug==2.0.3
jinja2==3.0.3
itsdangerous==2.0.1
transformers==4.30.2
torch==2.0.1
sentencepiece==0.1.99
protobuf==3.20.3
python-dotenv==1.0.0
numpy==1.24.3
huggingface-hub==0.16.4
acceleate==0.20.3